+++ src/low_level/mod.rs
@@ -16,7 +16,7 @@ mod signal_details;
 
 pub use signal_hook_registry::{register, unregister};
 
-pub use self::signal_details::signal_name;
+pub use self::signal_details::{emulate_default_handler, signal_name};
 
 /// The usual raise, just the safe wrapper around it.
 ///